requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
django-filter==23.5
django-prometheus==2.3.1
sentry-sdk==1.39.1
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-storages==1.14.2
boto3==1.34.34
python-jose==3.3.0
django-extensions==3.2.3
django-debug-toolbar==4.3.0
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.1
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.1.0
whitenoise==6.6.0
requests==2.31.0
openai==1.12.0
daphne==4.0.0 